moving cause query out a level since the 12 loaders was not a good look
